forward/requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
cachetools==4.2.2
certifi==2022.6.15
Django==4.0.6
psycopg2==2.9.3
python-dotenv==0.20.0
python-telegram-bot==13.15
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.2
tornado==6.2
tzdata==2022.1
tzlocal==4.2
